Fix projectile app initialization
This change was neccessary because the previous form was causing a bug
where the projectile's position was not being updated correctly, thus
causing the program to hang. The position was correct within the
underlying array, but the program was reading an incorrect position from
the reference variables. They never reflected the changes made to the
array. More investigation is needed to determine the underlying behavior
causing this bug.