Refactor Point, Vector & Color components
The components were changed from a reference to an array element into a
function that returns a reference to the same element. This was done
becuase it was mistakenly thought that the reference member was only
used by the compiler and occupied no space in memory. Upon inspection
and with the use of sizeof, it was found that each referenece did cause
the object to be larger than expected. It effectively doubles the amount
of memory needed for each class. For example a Point, which consists of
3 coordinates (doubles) & a reference to each coordinate, required 4
bytes per coordinate and 4 bytes per reference, for a total of 24 bytes.
The function returning a reference does not add to the object size.