Add constexpr Tuple ctor
It was found that the default construction of Tuple would cause the
underlying array to sometimes be non-zeroed. The Matrix-Tuple
multiplication operator default constructs a Tuple and assumes that it
is empty, which will sometimes work by chance but fail at other times.
Thus, the default constructor was changed to explicitly zero out the
array and constexpr was required due to the constexpr size member.